# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sink Company Inventory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type SysUserRole = (typeof SysUserRole)[keyof typeof SysUserRole];

export const SysUserRole = {
  admin: "admin",
  user: "user",
  auditor: "auditor",
} as const;

export interface SysUser {
  id: number;
  username: string;
  role: SysUserRole;
  assignedWarehouseId?: number | null;
  assignedWarehouseName?: string | null;
}

export interface AuthResponse {
  user: SysUser;
}

export interface MeResponse {
  user?: SysUser | null;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
  auditor: "auditor",
} as const;

export interface CreateUserBody {
  username: string;
  password: string;
  role: CreateUserBodyRole;
  assignedWarehouseId?: number | null;
}

export interface Warehouse {
  id: number;
  name: string;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  code: string;
  buyPrice: number;
  sellPrice: number;
  quantity: number;
  warehouseId: number;
  warehouseName?: string | null;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  code: string;
  buyPrice: number;
  sellPrice: number;
  quantity: number;
  warehouseId?: number;
}

export interface UpdateProductBody {
  name?: string;
  code?: string;
  buyPrice?: number;
  sellPrice?: number;
  quantity?: number;
}

export interface BulkResult {
  created: number;
  skipped: number;
}

export type MovementType = (typeof MovementType)[keyof typeof MovementType];

export const MovementType = {
  in: "in",
  out: "out",
} as const;

export interface Movement {
  id: number;
  type: MovementType;
  productCode: string;
  productName: string;
  quantity: number;
  price: number;
  total: number;
  warehouseId: number;
  warehouseName?: string | null;
  createdAt: string;
}

export type CreateMovementBodyType =
  (typeof CreateMovementBodyType)[keyof typeof CreateMovementBodyType];

export const CreateMovementBodyType = {
  in: "in",
  out: "out",
} as const;

export interface CreateMovementBody {
  type: CreateMovementBodyType;
  productCode: string;
  quantity: number;
  price: number;
  warehouseId?: number;
}

export interface InvoiceItem {
  productCode: string;
  productName: string;
  quantity: number;
  price: number;
  total: number;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  customerName: string;
  items: InvoiceItem[];
  total: number;
  status: string;
  warehouseId: number;
  warehouseName?: string | null;
  createdAt: string;
}

export type CreateInvoiceBodyItemsItem = {
  productCode: string;
  quantity: number;
  price: number;
};

export interface CreateInvoiceBody {
  customerName: string;
  warehouseId?: number;
  items: CreateInvoiceBodyItemsItem[];
}

export interface AuditLog {
  id: number;
  action: string;
  detail: string;
  username: string;
  createdAt: string;
}

export type DashboardChartItem = {
  date: string;
  in: number;
  out: number;
};

export interface Dashboard {
  totalProducts: number;
  totalQuantity: number;
  totalSales: number;
  profit: number;
  recordCount: number;
  recordLimit: number;
  lowStock: Product[];
  recentMovements: Movement[];
  chart: DashboardChartItem[];
}

export interface Report {
  totalIn: number;
  totalOut: number;
  totalQuantity: number;
  profit: number;
  movements: Movement[];
}

export type GetDashboardParams = {
  warehouseId?: number;
};

export type ListProductsParams = {
  warehouseId?: number;
  search?: string;
};

export type BulkCreateProductsBody = {
  warehouseId?: number;
  products: CreateProductBody[];
};

export type ListMovementsParams = {
  warehouseId?: number;
  type?: ListMovementsType;
  from?: string;
  to?: string;
  productCode?: string;
  limit?: number;
};

export type ListMovementsType =
  (typeof ListMovementsType)[keyof typeof ListMovementsType];

export const ListMovementsType = {
  in: "in",
  out: "out",
  all: "all",
} as const;

export type ListInvoicesParams = {
  warehouseId?: number;
};

export type CreateWarehouseBody = {
  name: string;
};

export type GetSettings200 = { [key: string]: string };

export type UpdateSettingsBody = { [key: string]: string };

export type ChangeUsernameBody = {
  username: string;
};

export type ChangePasswordBody = {
  password: string;
};

export type ListLogsParams = {
  limit?: number;
};

export type GetReportParams = {
  warehouseId?: number;
  from?: string;
  to?: string;
  productCode?: string;
};