# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sink Company Inventory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
...
  /movements:
    get:
      operationId: listMovements
      tags: [movements]
      parameters:
        - in: query
          name: warehouseId
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string, enum: [in, out, all] }
        - in: query
          name: from
          schema: { type: string }
        - in: query
          name: to
          schema: { type: string }
        - in: query
          name: productCode
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Movement" }
    post:
      operationId: createMovement
      tags: [movements]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMovementBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Movement" }
  /movements/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    delete:
      operationId: deleteMovement
      tags: [movements]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      parameters:
        - in: query
          name: warehouseId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoiceBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getInvoice
      tags: [invoices]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /account/username:
    put:
      operationId: changeUsername
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /account/password:
    put:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    SysUser:
      type: object
      required: [id, username, role]
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: string, enum: [admin, user, auditor] }
        assignedWarehouseId: { type: integer, nullable: true }
        assignedWarehouseName: { type: string, nullable: true }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/SysUser" }
...
    Movement:
      type: object
      required: [id, type, productCode, productName, quantity, price, total, warehouseId, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [in, out] }
        productCode: { type: string }
        productName: { type: string }
        quantity: { type: integer }
        price: { type: number }
        total: { type: number }
        warehouseId: { type: integer }
        warehouseName: { type: string, nullable: true }
        createdAt: { type: string }
    CreateMovementBody:
      type: object
      required: [type, productCode, quantity, price]
      properties:
        type: { type: string, enum: [in, out] }
        productCode: { type: string }
        quantity: { type: integer }
        price: { type: number }
        warehouseId: { type: integer }
    InvoiceItem:
      type: object
      required: [productCode, productName, quantity, price, total]
      properties:
        productCode: { type: string }
        productName: { type: string }
        quantity: { type: integer }
        price: { type: number }
        total: { type: number }
    Invoice:
      type: object
      required: [id, invoiceNumber, customerName, items, total, status, warehouseId, createdAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        customerName: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        total: { type: number }
        status: { type: string }
        warehouseId: { type: integer }
        warehouseName: { type: string, nullable: true }
        createdAt: { type: string }
    CreateInvoiceBody:
      type: object
      required: [customerName, items]
      properties:
        customerName: { type: string }
        warehouseId: { type: integer }
        items:
          type: array
          items:
            type: object
            required: [productCode, quantity, price]
            properties:
              productCode: { type: string }
              quantity: { type: integer }
              price: { type: number }
...