# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sink Company Inventory API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvoiceBodyItemsItem } from "./createInvoiceBodyItemsItem";

export interface CreateInvoiceBody {
  customerName: string;
  warehouseId?: number;
  items: CreateInvoiceBodyItemsItem[];
}