# lib/api-zod/src/generated/types/sysUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sink Company Inventory API
 * OpenAPI spec version: 0.1.0
 */
import type { SysUserRole } from "./sysUserRole";

export interface SysUser {
  id: number;
  username: string;
  role: SysUserRole;
  assignedWarehouseId?: number | null;
  assignedWarehouseName?: string | null;
}